Potential workaround for DGC in high level query on cluster.  I have introduced a ThickFuture. This class is used to turn a fast range count request into a sync RPC call by ensuring that we return a Future which embodies the output, rather than a proxy to communicate back to the DS to obtain the outcome.  This would appear to make sense for the low latency fast range count requests as we serve a LOT of them.
This commit also includes a small change to ListServices to report the nodes which are running the Jini service registrars.

The RangeCountProcedure was modified to self-report its [exact] and [deleted] attributes.

The DataServer was modified to use the ThickFuture for this operation.

The DataService#wrapFuture() method was removed.  It was not used anywhere in the code base (dead code).

https://sourceforge.net/apps/trac/bigdata/ticket/437
https://sourceforge.net/apps/trac/bigdata/ticket/433


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@5830 8f7bc0f5-282e-0410-95e3-8d296e9bb460