Ported r5824 and r5828 to the 1.0.x maintenance branch.  These change sets include the support for striped locks in the batched hard reference queue backing the concurrent hash map used by the termCache.
Note: The TermId.cache mechanism also needs to be defeated.  This was
done in r5829 in the BIGDATA_RELEASE_1_1_0 maintenance branch.
However, the IV class hierarchy is significantly different for 1.1.x
so this change set needs to be reimplemented for 1.0.x rather than
simply imported.  An ITermCache, TermCache, and IV.clone(boolean) were
all introduced to defeat this memory leak.

@see http://sourceforge.net/apps/trac/bigdata/ticket/437 (termCache and striped locks memory leak)


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_0_0@5836 8f7bc0f5-282e-0410-95e3-8d296e9bb460