Port of r5829 to the 1.0.x maintenance branch.
Encapsulate terms cache with interface and ensure IV.cache reference
is cleared when adding to the map.

- ITermCache
- TermCache
- LexiconRelation#termCache, #termCacheFactory

All IVs must implement clone(boolean).

- TestIVCache
- IV.dropCache(), IV.clone(boolean)
- AbstractIV#setValue() => final
- AbstractIV#dropCache() => removed.
- AbstractIV#flags => protected.
- TermId#clone(boolean).
- TermId#setValue(...) => removed (overrode AbstractIV but added nothing)
- NullIV#clone(boolean) returns self.
- DummyIV#clone(boolean) returns self.
- TestEncodeDecodeKeys: implemented clone(boolean) in private IV impl.
- TestEncodeDecodeKeys: imported logic to test clone(boolean).
- ExtensionIV#clone(boolean)
- SidIV#clone(boolean) & removed unused logger.
- NumericBNodeIV#clone(boolean)
- UUIDBNodeIV#clone(boolean)
- UUIDLiteralIV#clone(boolean)
- XSDBooleanIV#clone(boolean)
- XSDByteIV#clone(boolean)
- WrappedIV#clone(boolean): <<<=== Review with MikeP.

Note: No changes were required in order to encapsulate access to the
termsCache using ITermCache.  While it is used by various index
procedures in the 1.0.x maintenance branch, all of those references
are from inner classes which are directly accessing the termsCache
reference on the LexiconRelation.  (This was changed in the 1.1.x
branch, which now has static helper classes for those tasks.)

@see https://sourceforge.net/apps/trac/bigdata/ticket/437 (termsCache and thread-local memory leak)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_0_0@5837 8f7bc0f5-282e-0410-95e3-8d296e9bb460