It turns out that JDK 6 does not support NIO for multicast (support for this was added in JDK 7).  This means that the GangliaListener blocks and does not notice an interrupt.
I modified the GangliaListener to test for an interrupt.  It will now exit once interrupted, but it can be blocked waiting for an interrupt if there are no packets available.

I modified the GangliaService shutdown to push through a mock packet in order to ensure that the listener terminates quickly.

I added a unit test for shutdown of the GangliaService.  I should do one specifically for the listener as well.

New bigdata-ganglia jar.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@5868 8f7bc0f5-282e-0410-95e3-8d296e9bb460