Potential fix for [1].
This stack trace was observed on the clustered database.  Because it submits operations in parallel against shards it is possible that an interrupt could be thrown out of ClientIndexView which might not appear out of a BTree object. 

Regardless, it seems that the interrupt SHOULD be handled as an indication that processing should halt.  The means to halt processing is to close the buffer (aka the sink) in AbstractChunkedResolver and break out of the loop consuming chunks from the source.  This should have the semantics of interrupting the operation.

My main question on a fix for this is whether we should be using BlockingBuffer#close() or BlockingBuffer#abort(Throwable).  I think that close() will be fine.  It will essentially eat the interrupt, but it will halt processing normally, which is what we want.
 
[1] https://sourceforge.net/apps/trac/bigdata/ticket/460 (Uncaught interrupt resolving RDF terms)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@5898 8f7bc0f5-282e-0410-95e3-8d296e9bb460