Fix for KeyAfterPartitionException / KeyBeforePartitionException on cluster.
The exception appears inside of a pipeline join on a data service.  The predicate was mapped onto an appropriate shard.  However, the problem would appear to be that the predicate spans the shard boundary.  This results in a fromKey/toKey combination for the range count for the predicate which lies outside of the shard's fromKey/toKey.

My inclination is that the rangeCheck is inappropriate in this circumstance.  The rangeCount() request was perfectly fine, and it is fine that the predicate spans the shard boundary since the predicate should have been mapped across all shards which it spans.

While we could "fix" this by imposing the shard's boundaries onto the rangeCount in FusedView, I am not sure that this adds anything since we are just doing more work to defeat the key range check.  Instead, my inclination is to simply disable the key range check for AbstractBTree.

To that end, I have refactored the logic for AbstractBTree#rangeCheck() into RangeCheckUtility so it can be reused later if desired.  I have disabled AbstractBTree#rangeCheck() - it now always returns true.  The KeyAfterPartitionException and KeyBeforePartitionException classes were modified so they can be used with an ISeparatorKeys instance, and not just a LocalPartitionMetadata object.  This makes them compatible with PartitionLocator as well.  I also modified AbstractBTreeTupleCursor constructor which had two invocations of rangeCheck() which were not in a Java assert to use assert.  This is consistent with its declared throwables.  I have verified that the entire B+Tree test suite passes with these changes.

This should also resolve [1], which was also diagnosed as the rangeCheck operator failing to appropriate constrain the key range before it hit this check.

[1] https://sourceforge.net/apps/trac/bigdata/ticket/438 (KeyBeforePartitionException during closure).
[2] https://sourceforge.net/apps/trac/bigdata/ticket/461 (KeyAfterPartitionException / KeyBeforePartitionException on cluster)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@5904 8f7bc0f5-282e-0410-95e3-8d296e9bb460