Enables IndexMetadata recycling (this code path is nearly never used as only scale-out tends to update the IndexMetadata object after the index has been created).
Adds TestBTreeRecycle with a test suite focused on the behavior of writeCheckpoint2().  This covers Checkpoint, root addr, IndexMetadata, and bloom filter recycling.  There are two versions of the test. One with and one without a bloom filter.  The test suite does not examine how nodes and leaves are recycled in a large index, just an index having a single root leaf.

@see http://sourceforge.net/apps/trac/bigdata/ticket/473 (PhysicalAddressResolutionException after reopen using RWStore and recycler)

@see http://sourceforge.net/apps/trac/bigdata/ticket/476 (Test suite for writeCheckpoint() and recycling for BTree/HTree)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_0_0@5977 8f7bc0f5-282e-0410-95e3-8d296e9bb460