Modified the binding set encoder/decoder for the wire to write the namespace of the lexicon relation into the encoded record when it visits the first IVCache association.  The namespace is used to obtain the BigdataValueFactory and the BigdataValueSerializer.  The namespace will remain undiscovered (and uncommunicated) unless there is at least one IVCache association in the encoded solutions.
Added a stream-oriented API to encoder/decoder.

Unit test for stream-oriented encoder/decoder.

Replaced the serialization in the ThickChunkMessage with the new wire
encoder/decoder.

Implemented externalizable for ThickChunkMessage, which replaces its
POJO serialization.  The new format is versioned.

Adapted the FederatedQueryEngine test suite to run with
LocalChunkMessage in order to avoid an RDF dependency in the data
model.  I added an annotation to the FederatedQueryEngine for this
purpose.

@see https://sourceforge.net/apps/trac/bigdata/ticket/475 (Optimize serialization for query messages on cluster)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@6038 8f7bc0f5-282e-0410-95e3-8d296e9bb460