Implemented support for VIRTUAL GRAPHS.
Modified BatchRDFValueResolver to cache the Value on the IV for
resolved IVs.

Added unit test for VIRTUAL GRAPH extention to SPARQL at the AST layer.

Added unit test for evaluation of VIRTUAL GRAPH extension. (In fact,
this is hardly required since the translation is identical to the
translation of FROM and FROM NAMED).

There is some tricky business concerning the absence of a named graph
or default graph data set.  I followed the notes in DataSetSummary on
this point:

{{{
        /*
         * Note: Per DAWG tests graph-02 and graph-04, a query against an empty
         * default graph collection or an empty named graph collection should
         * be constrained to NO graphs.  This is different from the case where
         * the dataset is simply not specified, which is interpreted as having
         * no constraint on the visited graphs.  If you uncomment the next two
         * lines, both graph-02 and graph-04 in the TCK will fail.
         */
}}}

@see https://sourceforge.net/apps/trac/bigdata/ticket/493 (Support Virtual Graphs)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@6059 8f7bc0f5-282e-0410-95e3-8d296e9bb460