Additional work on SPARQL 1.1 Federated Query.
Added the UUID of the running query to the SolutionLog for the solutions drained from the query.

Added logic to throw an UnsupportedOperationException in the SearchServiceFactory if there are solutions being vectored into the search operator. The search services does not currently handle this case because it always winds up being invoked with an empty solution set (in a named subquery).  However, it should not silently ignore any solutions which are vectored in.

Modified the RemoteServiceCallImpl to use the factory pattern for the IRemoteSparqlQueryBuilder.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@6071 8f7bc0f5-282e-0410-95e3-8d296e9bb460