AbstractQueryEngineTestCase: Added logic to report the solutions which were not consumed.
RemoteSparql10Builder and RemoteSparql11Builder: Added a hack to support testEmptyServiceGraph.  Since there are no variables in the service clause, it was sending an invalid SELECT expression.  It is now hacked to send a "*" instead.

AbstractHashJoinUtilityTestCase: Added detailed example for openrdf service04.  As far as I can tell, the test case for this service query is wrong in openrdf.

@see http://sourceforge.net/apps/trac/bigdata/ticket/449 (SPARQL 1.1 Federation extension)


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@6089 8f7bc0f5-282e-0410-95e3-8d296e9bb460