Basic Federated Query is done [2].  There are still some open issues related to Federated Query, but they are optimization and edge cases.
I've added some variant SERVICE tests (variants of the openrdf tests) into the Federated Query test suite.

I have implemented a hack for [1] (blank nodes inside of the SERVICE graph pattern).   The hack covers the openrdf test10 problem.  It is in AST2BOpUtility#addServiceCall() and is documented there for what it is.  The easy workaround for this is to use variablers in the SPARQL instead of blank nodes.

Various javadoc notes on static analysis for exogenous variables [3] and for pruning variables [4].  These two issues are somewhat related.  We should really annotate the AST with the set of variables which are in scope and in use.  It seems like this would be best done as its own optimizer pass over the AST.

[1] https://sourceforge.net/apps/trac/bigdata/ticket/510 (Blank nodes in SERVICE graph patterns)
[2] https://sourceforge.net/apps/trac/bigdata/ticket/449 (Basic Federated Query).
[3] https://sourceforge.net/apps/trac/bigdata/ticket/412 (Static analysis of exogenous variables)
[4] https://sourceforge.net/apps/trac/bigdata/ticket/368 (Pruning variable bindings during query evaluation)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@6103 8f7bc0f5-282e-0410-95e3-8d296e9bb460