Actually, the abstract class IVValueExpression is a reasonable base class for the "function bops".  Pretty much everything appears to extend that abstract class.  All of the IFactory implementations and all of the "function bops" would have to be modified to pass the NAMESPACE of the lexicon relation into this abstract base class.
All classes which extend AbstractLiteralBOp already have this pattern.  They pass through the NAMESPACE annotation and have access to the BigdataValueFactory through a double-checked locking caching mechanism based on a volatile transient property.  That API could be easily extended to tunnel the IBindingSet to gain access to the BOpContext attached to the IBindingSet, obtain the IIndexManager, resolve the LexiconRelation, and expose a cached LexiconConfiguration reference.

The abstract class AbstractAccessPathOp has a RELATION_NAME[] which conveys similar information.  That should really be changed to a RELATION_NAME String and could even be the same NAMESPACE annotation which is used by the function bops. 

AbstractAccessPathOp also declares the TIMESTAMP of the view.  We really need that annotation to recover the right view of the LexiconRelation.  However, the ILexiconConfiguration metadata is immutable so it is Ok to use the last committed time for that view.

I have relayered the bops dealing with IVs from AbstractLiteralBOp.  This required fixing several factories in the FunctionRegistry and several unit tests.  I had to introduce a new base class (AbstractIVValueExpressionBOP2) which extends AbstractIVValueExpressionBOp but (unlike AbstractLiteralBOp) does NOT implement the INeedsMaterialization interface.

I have added a ContextBindingSet which delegates to a backing IBindingSet and also exposes the BOpContext.  This is automatically wrapped around binding sets which are not already instances of that class.  However, I have not yet chased down the code paths which create new binding sets so this information could be lost when copying binding sets (projection), joining binding sets, etc.  The delegate wrapper pattern is disabled until I chase all of that down and AbstractLiteralBOp#getLexiconConfiguration() will not work until it is enabled.

I have (hopefully) avoided needing to pass through the lexicon namespace into the SPARQLConstraint and FilterNode, both of which wrap the value expression with an EBVBop.  I think that EBVBOp can get along without this information.  I added an isLexiconNamespaceRequired() protected method to AbstractLiteralBOp.  It will only throw an exception if the NAMESPACE annotation is not present when that method returns true (which is the default behavior).

The test suites for the AST and SPARQL are green.  I still do not have the BOPContext exposed.

The following bops look like they should subclass AbstractLiteralBOp:

- DatatypeBOp
- DateBOp
- FuncBOp
- IfBOp
- LangBOp
- MathBOp
- NumericBOp
- RandBOp
- StrBOp
- XsdStrBOp

The unit tests for the SPARQL to AST translation, AST evaluation, SPARQL evaluation, the LocalTripleStore, the NSS in quads mode, and the BigdataSail in quads mode are all green.

It *is* possible that there will be some performance regressions associated with this commit.  We no annotation most value expressions with the Lexicon namespace, which means more LinkedHashMaps for that annotation.  Also, there may be subtle interactions with the value expression operator implementations which were not caught by the unit tests.

I have overridden ContextBindingSet methods which involve clone to return a clone of itself, not the delegate.  This is still not enabled.

@see https://sourceforge.net/apps/trac/bigdata/ticket/513 (Expose the LexiconConfiguration to function bops)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@6122 8f7bc0f5-282e-0410-95e3-8d296e9bb460