Modified the RemoteRepository to use an pattern for the connect options which can be overridden.
Integrated access to the ClientConnectionManager into the QueryEngine.  The ClientConnectionManager is lazily obtained from a factory and is shutdown() when the QueryEngine is shutdown.

Refactored the RemoteServiceCallImpl to use the RemoteRepository, which is now backed by the http components package.  This required a change in the ServiceRegistry API.

The RemoteRepository now uses more sophisticated defaults for the Accept header.  There is a new class to support this (AcceptHeaderFactory).


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@6195 8f7bc0f5-282e-0410-95e3-8d296e9bb460