I've dug into this a bit further.  If you look at the top-level WHERE clause, you can see: {{{ SELECT ( ... )   JoinGroupNode {     INCLUDE -subSelect-1 JOIN ON ()     INCLUDE -subSelect-2 JOIN ON ()     INCLUDE -subSelect-3 JOIN ON (VarNode(country),VarNode(product))   } }}}
Clearly, we do not have shared join variables for these INCLUDEs / sub-selects.

There is a bug in the logic to recognized when there are INCLUDEs having the same join variables.  The code was relying on a non-empty joinVars collection to indicate that it had processed the 2nd of the includes.  However, since there were no join variables for that include either it was willing to let the 3rd include into the set of required includes.  While the code is willing to generate a hash index for the first include, it does not attempt to do that for the 2nd+ includes.  

The simple fix is to this logic so it will reject this query as a candidate for a merge join.

However, we should also look more closely at how we are attaching join variables to sub-selects.  For this query, the different sub-selects have <code>country</code> as a shared projected variable.  So, we SHOULD be doing a hash join on <code>country</code>.  In fact, the code is falling back to a cross product hash join which is very non-optimal.

If we had determined during static analysis that each sub-select would definitely bind <code>country</code> then we could run the merge join since they would have a shared variable.  Failing to get this join variable right is probably a major performance hit.

After all that the predicted result for openrdf differs at the 4th decimal place. I have therefore filtered out this test from the openrdf TCK.

<pre>
Missing bindings: 
[product=http://www4.wiwiss.fu-berlin.de/bizer/bsbm/v01/instances/dataFromProducer2/Product63;
 nrOfReviews="3"^^<http://www.w3.org/2001/XMLSchema#integer>;
 avgPrice="4207.426"^^<http://www.w3.org/2001/XMLSchema#float>;
 country=http://downlode.org/rdf/iso-3166/countries#RU]
====================================================
Unexpected bindings: 
[country=http://downlode.org/rdf/iso-3166/countries#RU;
 product=http://www4.wiwiss.fu-berlin.de/bizer/bsbm/v01/instances/dataFromProducer2/Product63;
 nrOfReviews="3"^^<http://www.w3.org/2001/XMLSchema#integer>;
 avgPrice="4207.4263"^^<http://www.w3.org/2001/XMLSchema#float>]
</pre>


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@6218 8f7bc0f5-282e-0410-95e3-8d296e9bb460