The FILTER NOT EXISTS issue is fixed.  The problem was the ordering of the children in the graph pattern group.  The fix was to the ASTJoinOrderByTypeOptimizer.  We are modeling the NOT EXISTS and EXISTS graph patterns as ASK subqueries.  It was treating those subqueries as required joins and ordering them early in some cases.
I have modified the ASTJoinOrderByTypeOptimizer to order the ASK subqueries after the required joins.  

While this change fixes this query, it is possible that we still could get bad join orderings when the variables used by the filter are only bound by OPTIONAL joins. It is also possible that we could run the ASK subquery for FILTER (NOT) EXISTS earlier if the filter variables are bound by required joins. This is really identical to the join filter attachment problem. The problem in the AST is that both the ASK subquery and the FILTER are present. It seems that the best solution would be to attach the ASK subquery to the FILTER and then to run it immediately before the FILTER, letting the existing filter attachment logic decide where to place the filter. We would also have to make sure that the FILTER was never attached to a JOIN since the ASK subquery would have to be run before the FILTER was evaluated.

@see https://sourceforge.net/apps/trac/bigdata/ticket/515 (Query with two "FILTER NOT EXISTS" expressions returns no results)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_1_0@6227 8f7bc0f5-282e-0410-95e3-8d296e9bb460