This commit incorporates the memory manager journal mode (MemStrategy, which is enabled by BuffedMode.MemStore) [1].
It also incorporates some progress on the bigdata extension for SPARQL Update [2].  I am beginning to work through the integration to support this feature.  I have written code for writing solution sets onto the memory manager and reading them back.  That code has not yet been tested.  This functionality is currently disabled in QueryHints and in the test suite.

[1] https://sourceforge.net/apps/trac/bigdata/ticket/541 (MemoryManager Journal mode)
[2] https://sourceforge.net/apps/trac/bigdata/ticket/531 (SPARQL UPDATE for Solution Sets).

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6266 8f7bc0f5-282e-0410-95e3-8d296e9bb460