Modified the static analysis code to work with both named subquery roots (the named solution set is produced within the query) and pre-existing named solution sets (sparql cache).
Modified the ISolutionSetStats class to always provide the identified
constants and to also identify any variables which are materialized in
the named solution set (the variable binding always has its IVCache
association set when the variable is bound in a solution).  This
supports the static analysis of pre-existing named solution sets.

The SolutionSetStats class is no longer used in the main code base (it
still exists in the test suite).  This is because the streaming
iterator variant now computes all data which was computed by the
SolutionSetStats class but can also be applied to large solution sets
such as are emerging from hash joins without requiring all solutions
to be materialized in memory.

Bug fix to TestASTBottomUpOptimizer.  It was using BigdataValues in
the solutions flowing through the optimizer rather than IVs.

@see https://sourceforge.net/apps/trac/bigdata/ticket/531 (SPARQL UPDATE for solution sets)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6293 8f7bc0f5-282e-0410-95e3-8d296e9bb460