Identified a leak in the NanoSparqlServer test suite (TestNanoSparqlServerWithProxyIndexManager).  This has a method getJournal() which returns a transient mode Journal but fails to correctly close the journal at the end of the test suite.  This appears to be leaking direct buffers when using bloom filters and is probably leaking threads for the Journal as well.
This ticket might have an impact on load performance since there is now a bloom filter on each index up to ~ 2,000,000 triples / quads.  This will need to be tested by a benchmark.

In the meanwhile, I have modified SPORelation#getStatementIndexMetadata() to only enable the bloom filter on the SPO and SPOC indices.  

 https://sourceforge.net/apps/trac/bigdata/ticket/150 (Choosing the index for testing fully bound access paths based on index locality)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6358 8f7bc0f5-282e-0410-95e3-8d296e9bb460