I have finished relayering the Checkpoint, ICheckpointProtocol, ISolutionSet, and SolutionSetStream classes.  There is now a durable SolutionSetStream which can be backed by an IRWStrategy.  Once we resolve [1], we will be able to generalize this to any IRawStore.  The SparqlCache is still wired to a ConcurrentHashMap.  I have not yet derived versions of the ISolutionSetCache for the HTree and the BTree.  There is some remaining work to be done there concerning the integration with the Checkpoint record. Specifically, I need to establish a protocol to store the address of the ISolutionSetStats in the checkpoint record. Right now the ISolutionSetStats is hacked into the bloom filter addr since that is not used by the SolutionSetStream.  There is plenty of room in the Checkpoint record, we just need to establish a protocol to make this work.
There is now a top-level durable Stream.  SolutionSetStream extends this.  The Stream is based on the PSOutputStream mechanism but it implements the ICheckpointProtocol and hence can be stored in Name2Addr just like the BTree or HTree.  This also makes it compatible with a MemStore mode Journal for the SPARQL cache.

I still need to review the integration of the SparqlCache with the SolutionSetStream and provide appropriate layering of access (query scope => cache scope => database scope) through the resource locator protocol. There are notes on this above and also in the committed files.

[1] https://sourceforge.net/apps/trac/bigdata/ticket/555 (Support PSOutputStream/InputStream at IRawStore)

@see https://sourceforge.net/apps/trac/bigdata/ticket/524 (SPARQL Cache)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6372 8f7bc0f5-282e-0410-95e3-8d296e9bb460