The following three openrdf SPARQL UPDATE tests have been disabled since they violate the SPARQL UPDATE specification by having blank nodes in the DELETE template:
- testDeleteTransformedWhere
- testUpdateSequenceDeleteInsert
- testUpdateSequenceInsertDelete

I have also picked up several new SPARQL UPDATE tests from Sesame releases up through 2.6.6 (this is not a general dependency upgrade - I only looked at this one file).  The changes to the SPARQL update test suite have been applied to both the local and NSS versions of the test suite.

The UpdateExprBuilder was modified to forbid blank nodes in the DELETE clause template.  It also forbids blank nodes in a DELETE DATA clause.  The test suite for the UpdateExprBuilder was extended to verify correct rejection for blank nodes in those locations.

The AST2BOpUpdate class was modified to NOT replace blank nodes with wildcards.

The test now runs green.

@see https://sourceforge.net/apps/trac/bigdata/ticket/571 (DELETE/INSERT WHERE handling of blank nodes.)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6376 8f7bc0f5-282e-0410-95e3-8d296e9bb460