Simplified the AccessPath class by moving AccessPath#get(int index) into the test suite. It was only used by two of the backchaining expander test suites.
Removed IAccessPath#tupleIterator().  It is still implemented internally by AccessPath, but it is no longer exposed and it is not present on the other IAccessPath implementation classes.

Lifted out an ISimpleIndexAccess interface from ICheckpointProtocol with rangeCount(), scan(), and removeAll() methods.

Isolated a getAccessPath() method returning an IBindingSetAccessPath in the HTreeHashJoinOp and JVMHashJoinOp classes.  This currently uses BOpContext.getAccessPath(Relation,Predicate).  However, this should be the only method that we need to change in order to execute a hash join against a SolutionSetStream.

The next step is to test a change from modeling named subquery outputs as hash indices to modeling those outputs as streams. I will have to change the named solution set resolution mechanism (to be simpler) and the query plans (to build the hash index from the pipeline).  I will also need to do a performance assessment for this change, probably for BSBM and the govtrack data sets.

@see https://sourceforge.net/apps/trac/bigdata/ticket/531 (SPARQL UPDATE for SOLUTION SETS)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6400 8f7bc0f5-282e-0410-95e3-8d296e9bb460