I have changed the log4j version from 1.2.15 to 1.2.17 to work around a dependency inclusion problem in the earlier version of log4j.
I have renamed the ICU dependencies to have version numbers that are consistent with the version numbers used in maven repositories for ICU (4_8 => 4.8).

I have added a pom for the embedded bigdata mode based on one contributed by Gerjon of the Open Sahara project - many thanks to Gerjon for his input on this effort.

The dependency version information is now maintained in build.properties (for ant), pom.xml (for maven), and .classpath (for Eclipse).

The maven pom should work at this point, but only for people who want to build their own version of bigdata since there are no published releases of bigdata yet in our maven repository.  We will publish 1.2.2 in this fashion.

Note that I have not taken on some of the challenges of the "maven-branch".  There are still plenty of cycles in the project packages.

@see https://sourceforge.net/apps/trac/bigdata/ticket/168 (Mavenization of bigdata)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6402 8f7bc0f5-282e-0410-95e3-8d296e9bb460