The IRDFParserOptions interface was lifted out of RDFParserOptions.
The SPARQL grammar was modified to accept options that can be specified in the same position as the SILENT keyword and a test suite was written to verify the correct interpretation of this extension to the SPARQL UPDATE grammar.  For example, you can now specify:

{{{
LOAD verifyData=true ...
LOAD stopAtFirstError=false ...
etc.
}}}

The historical value for verifyData when the zero argument version of the RDFParserOptions construct was used was TRUE when it should have been the same as the value specified for RDFParserOptions.Options.DEFAULT_VERIFY_DATA.  This has been fixed. It now defaults to FALSE. This has the same semantics as the default for RDFParserBase.

Note: The default for stopAtFirstError is NOT the same as the default for RDFParserBase.  We should probably change this.

@see https://sourceforge.net/apps/trac/bigdata/ticket/591 (SPARQL UPDATE "LOAD" extensions)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6483 8f7bc0f5-282e-0410-95e3-8d296e9bb460