Modified AbstractJournal to synchronize the root blocks when the quorum meets.
Extracted code from RootBlockUtility into a static public method to chose the current root block.  This preserves the historical decision to choose rootblock1 when the commit counters are the same.

Journal was not logging txLog OPEN events for the WORMStrategy.

Modified DumpZookeeper to show the effective session timeout negotiated with the zk quorum.  Fixed bug where it was not flushing out the PrintWriter.

Modified the DefaultResourceLocator to not use the UNISOLATED view of the GRS when the caller was using a READ-COMMITED view.  This addressed some problems with the follower, which is read-only.  Modified the DefaultResourceLocator test suite since the read-committed view of the GRS is not actually visible until after a commit on a Journal.

Modified the NSS startup to handle startup in a read-only mode (as a quorum follower).

Modified Journal.getCounters() to not report the index counters for a read-only journal (this lead to an NPE since _name2addr is null for a read-only journal).

Modified the HAJournal configuration files to use a short (5s) timeout for both zookeeper and jini. This makes debugging easier.  I am not sure yet what values we will recommend for deployment.  One reason to raise these values is that a long GC pause could cause a session timeout or registrar heartbeat drop.

@see https://sourceforge.net/apps/trac/bigdata/ticket/530 (Journal HA)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6547 8f7bc0f5-282e-0410-95e3-8d296e9bb460