Added the concept of a logicalServiceId to the HAJournalServer. Unlike the federation, which creates logical service ids automatically, the logical service id for the HAJournal is configured explicitly for the HAJournalServer component.  I also removed the PIPELINE entry in the HAJournalServer configuration.  The quorum members now gather around the logicalServiceId. This means that it is possible to have multiple HA Journal quorums running in the same zookeeper + river environment.
Modified the HAJournalServer to use the standard JiniClientConfig.

Added support for the JoinManager.leaseTimeout to the sample
configuration files.  Note that you MUST override the
LeaseRenewalManager if you want to use a lease timeout LT 20s (the
round trip time assumption is 10s and the lease renewal manager is
practive, so if the lease duration is 20s and the expected round trip
time is 10s, it will renew every 10s).

The JiniClientConfig had a constructor parameter "className" that was no longer used.  I have updated the javadoc to reflect this fact and deprecated the PROPERTIES configuration option in JiniClientConfig.Options since it really belongs in the JiniClient class (it is used with the component name com.bigdata.service.JiniClient) and not the JiniClientConfig class.

https://sourceforge.net/apps/trac/bigdata/ticket/530 (Journal HA)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6565 8f7bc0f5-282e-0410-95e3-8d296e9bb460