The change sets for the tickets identified in this issue have been ported into branches/BIGDATA_BRANCH_1_2_1.  I have also brought across the unit test for the RWStore.reset() issue and verified that it (a) fails before the code change to reset() and (b) succeeds with that code change to reset().
The RWStore test suite is green with these changes.  I will commit and setup CI on the new branch in preparation for tagging a 1.2.2 release from the branch.

I have added release notes for 1.2.2 to the branch.  A copy of those release notes will also be included in the development branch, which is now slated for a 1.2.3 release.

@see https://sourceforge.net/apps/trac/bigdata/ticket/603 (Prepare critical maintenance release as branch of 1.2.1)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_1@6581 8f7bc0f5-282e-0410-95e3-8d296e9bb460