This commit addresses an issue identified yesterday with HA where it was possible for a delete entry to be entered into the write cache buffer between while the buffer was being replicated, leading to an inconsistency between the #of bytes the receiver was expecting and the #of bytes that are sent.  A live lock could result where the receiver was expecting more data than the sender was transmitting.  This issue was resolved by Martyn by introducing a "closeForWrites()" method onto the WriteCache that serves to prevent the append of "delete entries" to a WriteCache's internal buffer when the buffer is about to be transmitted.
The getHASendService() and getHAReceiveService() methods were removed from the QuorumPipeline interface.  These methods can not be exposed while still making the access to the underlying services safe with respect to write pipeline and quorum state changes.  Further, these methods were never used outside of QuorumPipelineImpl. Therefore, the methods are now private to that class.

QuourmPiplineImpl was further refactored to (a) use a lock when observing the fields on that class so the observations are atomic with respect to state changes; and (b) pulled out the inner Callable classes into private static classes for send() and receiveAndReplicate().  This latter change was made in support of the atomic observation of the member fields on QuorumPipelineImpl.  The code is not holding the lock across the send() and receiveAndReplicate() operations, but perhaps it should.  We still need to look at this class in more depth.

The HA versions of the write replication pipeline tests were enabled in CI, except those those tests pertaining to pipeline reorganization.  I have observed test failures and deadlocks in the pipeline reorganization tests. I do not yet know if those problems are in the test suite or in the code under test.

Assertion for member == null added to HAReceiveService.

The HA quorum tests are enabled for CI in this commit.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6611 8f7bc0f5-282e-0410-95e3-8d296e9bb460