Bug fix to AbstractJournal to go through a low level abort for the leader also when a quorum meets (corrects the previous commit).
Modified HALogWriter/Reader to support the alternating root block pattern.  100% binary compatible journal and HALog files established through resynchronization protocol.

Added the serviceId to the IHALogRequest and now checking it as part of the resynchronization protocol to make sure that we only accept the HALog replay messages that we requested.

https://sourceforge.net/apps/trac/bigdata/ticket/530 (Journal HA)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6671 8f7bc0f5-282e-0410-95e3-8d296e9bb460