Stubbed in support for disaster recovery based on replication of the leader's backing file plus replay of the HALogs.  This involved a change to some of the HA interfaces (IHALogRequest is now a subinterface of IHASyncRequest).  New methods were added to request the backing store from the leader, to do the appropriate raw reads on the backing store, and to replicate the data onto the pipeline.
The HAJournalServer does not yet process the received write cache blocks for a rebuild operation.  We need to provide an appropriate hook for the RWStore since it needs to handle these more like WORM writes than scattered writes.  Also, the HALog replay is not yet in place. 

Methods were also stubbed in to allow the leader to request the digest of the stores from the followers.  These methods have not been implemented yet.

@see https://sourceforge.net/apps/trac/bigdata/ticket/530 (Journal HA)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_2_0@6696 8f7bc0f5-282e-0410-95e3-8d296e9bb460